rename str helper file in a06
