+++ src/a06_bud_logic/_utils/str_a06.py
