+++ src/a06_bud_logic/test_bud_tool/test_bud_attr_exists.py
@@ -14,7 +14,7 @@ from src.a06_bud_logic.bud_tool import (
     bud_item_factunit_exists,
     bud_attr_exists,
 )
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
