+++ src/a08_bud_atom_logic/test_/test_atom_core.py
@@ -1,5 +1,5 @@
 from src.a03_group_logic.acct import acctunit_shop
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
