+++ src/a08_bud_atom_logic/test_/test_atom_row.py
@@ -1,5 +1,5 @@
 from src.a01_word_logic.road import create_road
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     bud_acctunit_str,
     bud_acct_membership_str,
     bud_itemunit_str,
