+++ src/a08_bud_atom_logic/test_/test_atom_sift_insert.py
@@ -1,7 +1,7 @@
 from src.a03_group_logic.group import awardlink_shop
 from src.a04_reason_logic.reason_item import reasonunit_shop, factunit_shop
 from src.a06_bud_logic.bud import budunit_shop
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     bud_acctunit_str,
     bud_acct_membership_str,
     bud_itemunit_str,
