+++ src/a08_bud_atom_logic/test_/test_atom_sift_update.py
@@ -6,7 +6,7 @@ from src.a06_bud_logic.bud_tool import (
     bud_item_reason_premiseunit_get_obj as premiseunit_get_obj,
     bud_item_factunit_get_obj,
 )
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
