+++ src/a08_bud_atom_logic/test_normal_models/test_normal_models.py
@@ -1,4 +1,4 @@
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
