+++ src/a09_pack_logic/test_delta/test_delta_legible_acctunit.py
@@ -1,4 +1,4 @@
-from src.a06_bud_logic._utils.a06_str_helpers import bud_acctunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import (
     atom_update,
     atom_insert,
