+++ src/a09_pack_logic/test_delta/test_delta_legible_item_healerlink.py
@@ -1,4 +1,4 @@
-from src.a06_bud_logic._utils.a06_str_helpers import bud_item_healerlink_str
+from src.a06_bud_logic._utils.str_a06 import bud_item_healerlink_str
 from src.a08_bud_atom_logic.atom_config import atom_insert, atom_delete, healer_name_str
 from src.a08_bud_atom_logic.atom import budatom_shop
 from src.a09_pack_logic.delta import buddelta_shop
