+++ src/a09_pack_logic/test_delta/test_delta_sift_delete.py
@@ -1,7 +1,7 @@
 from src.a06_bud_logic.bud import budunit_shop
 
 # from src.a06_bud_logic.bud_tool import pass
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     bud_acctunit_str,
     bud_acct_membership_str,
 )
