+++ src/a09_pack_logic/test_pack/test_pack_.py
@@ -2,7 +2,7 @@ from src.a00_data_toolbox.dict_toolbox import x_is_json
 from src.a02_finance_logic._utils.strs_a02 import owner_name_str, fisc_tag_str
 from src.a01_word_logic.road import get_default_fisc_tag as root_tag
 from src.a03_group_logic.acct import acctunit_shop
-from src.a06_bud_logic._utils.a06_str_helpers import bud_acctunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_acctunit_str
 from src.a06_bud_logic.bud import budunit_shop
 from src.a08_bud_atom_logic.atom import budatom_shop
 from src.a08_bud_atom_logic.atom_config import (
