+++ src/a10_bud_calc/test_/test_bud_calc_config.py
@@ -5,7 +5,7 @@ from src.a02_finance_logic._utils.strs_a02 import (
     owner_name_str,
     world_id_str,
 )
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
