+++ src/a15_fisc_logic/test_fisc_/test_journal_sqlstr_.py
@@ -1,5 +1,5 @@
 from src.a01_word_logic.road import create_road
-from src.a06_bud_logic._utils.a06_str_helpers import bud_item_factunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_item_factunit_str
 from src.a08_bud_atom_logic.atom_config import (
     fopen_str,
     atom_insert,
