+++ src/a17_idea_logic/test_idea/test_idea_format_builder.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.file_toolbox import save_json
 from src.a02_finance_logic._utils.strs_a02 import owner_name_str, fisc_tag_str
-from src.a06_bud_logic._utils.a06_str_helpers import bud_itemunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_itemunit_str
 
 # from src.a06_bud_logic._utils.str_helpers import
 from src.a08_bud_atom_logic.atom_config import (
@@ -15,7 +15,7 @@ from src.a17_idea_logic.idea_config import (
     dimens_str,
     attributes_str,
 )
-from src.a06_bud_logic._utils.a06_str_helpers import bud_itemunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_itemunit_str
 
 
 def create_dimens_idea_format_dict() -> dict:
