+++ src/a17_idea_logic/test_idea/test_idea_format_config.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.file_toolbox import get_dir_file_strs, create_path
 from src.a02_finance_logic._utils.strs_a02 import owner_name_str, fisc_tag_str
-from src.a06_bud_logic._utils.a06_str_helpers import bud_acctunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import (
     face_name_str,
     event_int_str,
