+++ src/a17_idea_logic/test_idea/test_idea_format_generate_idea.py
@@ -1,6 +1,6 @@
 from src.a05_item_logic.item import itemunit_shop
 from src.a06_bud_logic.bud import budunit_shop
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     bud_acctunit_str,
     bud_acct_membership_str,
     bud_itemunit_str,
