+++ src/a18_etl_toolbox/test_tran/test_etl_event_bud_csvs_to_pack_jsons.py
@@ -1,7 +1,7 @@
 from src.a00_data_toolbox.file_toolbox import open_file, save_file
 from src.a02_finance_logic._utils.strs_a02 import owner_name_str, fisc_tag_str
 from src.a06_bud_logic._utils.example_buds import from_list_get_active
-from src.a06_bud_logic._utils.a06_str_helpers import bud_acctunit_str, budunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_acctunit_str, budunit_str
 from src.a08_bud_atom_logic.atom_config import (
     face_name_str,
     acct_name_str,
