+++ src/a18_etl_toolbox/test_tran/test_idea_raw_tables2bud_raw_tables.py
@@ -1,7 +1,7 @@
 from src.a00_data_toolbox.file_toolbox import create_path, save_file
 from src.a00_data_toolbox.db_toolbox import db_table_exists, get_row_count
 from src.a02_finance_logic._utils.strs_a02 import owner_name_str, fisc_tag_str
-from src.a06_bud_logic._utils.a06_str_helpers import budunit_str, bud_acctunit_str
+from src.a06_bud_logic._utils.str_a06 import budunit_str, bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import (
     face_name_str,
     acct_name_str,
