+++ src/a19_world_logic/test_06_create_guts/test_world_etl06_02_bud_raw_tables_to_bud_agg_tables.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.db_toolbox import get_row_count
 from src.a02_finance_logic._utils.strs_a02 import owner_name_str, fisc_tag_str
-from src.a06_bud_logic._utils.a06_str_helpers import bud_acctunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import (
     face_name_str,
     acct_name_str,
