+++ src/a19_world_logic/test_06_create_guts/test_world_etl06_05_fisc_events_pack_jsons_to_budunit_json.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.file_toolbox import create_path, open_file, save_file
 from src.a06_bud_logic.bud import budunit_shop, get_from_json as budunit_get_from_json
-from src.a06_bud_logic._utils.a06_str_helpers import bud_acctunit_str
+from src.a06_bud_logic._utils.str_a06 import bud_acctunit_str
 from src.a08_bud_atom_logic.atom_config import (
     acct_name_str,
     credit_belief_str,
