+++ src/a20_lobby_db_toolbox/test_job/test__job_create_sqlstrs.py
@@ -6,7 +6,7 @@ from src.a00_data_toolbox.db_toolbox import (
     create_select_query,
 )
 from src.a02_finance_logic._utils.strs_a02 import fisc_tag_str, owner_name_str
-from src.a06_bud_logic._utils.a06_str_helpers import (
+from src.a06_bud_logic._utils.str_a06 import (
     budunit_str,
     bud_acctunit_str,
     bud_acct_membership_str,
